requirements.txt
Django==3.2.23
psycopg2-binary==2.9.9
gunicorn==21.2.0
djangorestframework==3.12.1
markdown==3.3.2
django-filter==2.4.0
django-crispy-forms==2.0.0
